+++ Sources/VelocityAdsGmaAdapter/VelocityInterstitialAdapterDelegate.swift
@@ -37,7 +37,7 @@ final class VelocityInterstitialAdapterDelegate: NSObject, VelocityInterstitialA
 
     func onAdLoaded(ad: any VelocityFullscreenAd) {
         guard let mediationAd else {
-            _ = completionHandler?(nil, VelocityAdsErrorMapper.adNotReady())
+            _ = completionHandler?(nil, VelocityAdsErrorMapper.adapterReleased())
             completionHandler = nil
             return
         }